Update init_bc_config_install_or_upgrade.php
The option to change the background color of Home, Log In, etc. when hovering over them was not included in the original listing.  The existing stylesheet color of #333333 is more than sufficient but if the owner changes the color scheme of the navbar there's no easy option to change this setting.

As to the other color changes, more sophisticated testing methods actually require a higher color contrast for "action" text like links, buttons, etc.

This reflects the changes needed to comply with those better defined rules.