Adding check for "Show Greeting" setting
This setting can be configured in the ZC Admin Backend under "Configuration > Layout Settings > Switch Customer Greeting > Show on Index" and Set to 0.

To keep the idea of accessibility in hand, if the HEADING_TITLE is not empty and the $show_welcome flag is set to true, add the <h1> under normal circumstances. If however either the flag is set to false or the HEADING_TITLE is empty, "show" the alternative accessibility version instead.